Enable RadioReference import of frequencies not within an agency
Some counties have frequencies associated without an intermediary
Agency. Previously, those frequencies could not be imported because
there was no UI available to support it.

This patch adds a "(Countywide)" option to the existing agency selection
box. When selected, the user can choose from the
categories/subcategories for the countywide frequencies using the
existing components.